Adding repositories

+++ src/main/java/com/marjorye/sima/models/Class_.java
@@ -10,14 +10,14 @@ public class Class_ {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    int id;
+    long id;
     String name;
     String shift;
     int year;
 
     public Class_(){}
 
-    public int getId() {
+    public long getId() {
         return id;
     }
